Minor improvement/bug fix in daemonization procedure
1. _exit parent process instead of returning from main(), since
we'd rather that destructors/atexit functions (if any) not be called.

2. Don't try to unlink PID file if fork() fails and there is no PID file.

+++ rdiscd.cpp
@@ -388,12 +388,14 @@ int main (int argc, char** argv)
 			pid_t		pid = fork();
 			if (pid == -1) {
 				std::perror("fork");
-				unlink(pid_file);
+				if (pid_file) {
+					unlink(pid_file);
+				}
 				return 1;
 			}
 			if (pid != 0) {
 				// exit the parent process
-				return 0;
+				_exit(0);
 			}
 			setsid();
